pbio: fix poll stopping after 65.535 seconds
Apparently the compiler needs some help when subtracting two uint16_t to prevent it from treating the result as an int.

+++ lib/pbio/drv/move_hub/ioport.c
@@ -159,7 +159,7 @@ static void poll_dcm(ioport_t port, uint16_t now) {
     uint8_t gpio_input;
 
     // wait at least 2ms between steps
-    if (now - data.prev_msec < 2) {
+    if ((uint16_t)(now - data.prev_msec) < 2) {
         return;
     }
